+++ src/pages/home/report/ReportActionsView.js
@@ -1,27 +1,18 @@
 import React from 'react';
 import {
-    View,
     Keyboard,
     AppState,
-    ActivityIndicator,
 } from 'react-native';
 import {withOnyx} from 'react-native-onyx';
 import PropTypes from 'prop-types';
 import _ from 'underscore';
 import lodashGet from 'lodash/get';
-import Clipboard from '../../../libs/Clipboard';
 import * as Report from '../../../libs/actions/Report';
-import KeyboardShortcut from '../../../libs/KeyboardShortcut';
-import SelectionScraper from '../../../libs/SelectionScraper';
-import ReportActionItem from './ReportActionItem';
-import styles from '../../../styles/styles';
 import reportActionPropTypes from './reportActionPropTypes';
-import InvertedFlatList from '../../../components/InvertedFlatList';
 import * as CollectionUtils from '../../../libs/CollectionUtils';
 import Visibility from '../../../libs/Visibility';
 import Timing from '../../../libs/actions/Timing';
 import CONST from '../../../CONST';
-import themeColors from '../../../styles/themes/default';
 import compose from '../../../libs/compose';
 import withWindowDimensions, {windowDimensionsPropTypes} from '../../../components/withWindowDimensions';
 import withDrawerState, {withDrawerPropTypes} from '../../../components/withDrawerState';
@@ -30,16 +21,16 @@ import withLocalize, {withLocalizePropTypes} from '../../../components/withLocal
 import ReportActionComposeFocusManager from '../../../libs/ReportActionComposeFocusManager';
 import * as ReportActionContextMenu from './ContextMenu/ReportActionContextMenu';
 import PopoverReportActionContextMenu from './ContextMenu/PopoverReportActionContextMenu';
-import variables from '../../../styles/variables';
 import Performance from '../../../libs/Performance';
 import * as ReportUtils from '../../../libs/reportUtils';
 import ONYXKEYS from '../../../ONYXKEYS';
-import {withNetwork, withPersonalDetails} from '../../../components/OnyxProvider';
-import participantPropTypes from '../../../components/participantPropTypes';
-import EmojiPicker from '../../../components/EmojiPicker';
+import {withNetwork} from '../../../components/OnyxProvider';
 import * as EmojiPickerAction from '../../../libs/actions/EmojiPickerAction';
 import FloatingMessageCounter from './FloatingMessageCounter';
 import networkPropTypes from '../../../components/networkPropTypes';
+import ReportActionsList from './ReportActionsList';
+import CopySelectionHelper from '../../../components/CopySelectionHelper';
+import EmojiPicker from '../../../components/EmojiPicker';
 
 const propTypes = {
     /** The ID of the report actions will be created for */
@@ -77,9 +68,6 @@ const propTypes = {
     /** Are we waiting for more report data? */
     isLoadingReportData: PropTypes.bool,
 
-    /** Personal details of all the users */
-    personalDetails: PropTypes.objectOf(participantPropTypes),
-
     /** Information about the network */
     network: networkPropTypes.isRequired,
 
@@ -98,19 +86,12 @@ const defaultProps = {
     session: {},
     isLoadingReportActions: false,
     isLoadingReportData: false,
-    personalDetails: {},
 };
 
 class ReportActionsView extends React.Component {
     constructor(props) {
         super(props);
 
-        this.renderItem = this.renderItem.bind(this);
-        this.renderCell = this.renderCell.bind(this);
-        this.scrollToBottomAndUpdateLastRead = this.scrollToBottomAndUpdateLastRead.bind(this);
-        this.onVisibilityChange = this.onVisibilityChange.bind(this);
-        this.recordTimeToMeasureItemLayout = this.recordTimeToMeasureItemLayout.bind(this);
-        this.loadMoreChats = this.loadMoreChats.bind(this);
         this.sortedReportActions = [];
         this.appStateChangeListener = null;
 
@@ -124,7 +105,10 @@ class ReportActionsView extends React.Component {
         this.currentScrollOffset = 0;
         this.updateSortedReportActions(props.reportActions);
         this.updateMostRecentIOUReportActionNumber(props.reportActions);
-        this.keyExtractor = this.keyExtractor.bind(this);
+
+        this.scrollToBottomAndUpdateLastRead = this.scrollToBottomAndUpdateLastRead.bind(this);
+        this.recordTimeToMeasureItemLayout = this.recordTimeToMeasureItemLayout.bind(this);
+        this.loadMoreChats = this.loadMoreChats.bind(this);
         this.trackScroll = this.trackScroll.bind(this);
         this.showFloatingMessageCounter = this.showFloatingMessageCounter.bind(this);
         this.hideFloatingMessageCounter = this.hideFloatingMessageCounter.bind(this);
@@ -135,7 +119,13 @@ class ReportActionsView extends React.Component {
     }
 
     componentDidMount() {
-        this.appStateChangeListener = AppState.addEventListener('change', this.onVisibilityChange);
+        this.appStateChangeListener = AppState.addEventListener('change', () => {
+            if (!Visibility.isVisible() || this.props.isDrawerOpen) {
+                return;
+            }
+
+            Report.updateLastReadActionID(this.props.reportID);
+        });
 
         // If the reportID is not found then we have either not loaded this chat or the user is unable to access it.
         // We will attempt to fetch it and redirect if still not accessible.
@@ -155,15 +145,6 @@ class ReportActionsView extends React.Component {
         }
 
         this.fetchData();
-
-        const copyShortcutConfig = CONST.KEYBOARD_SHORTCUTS.COPY;
-        this.unsubscribeCopyShortcut = KeyboardShortcut.subscribe(
-            copyShortcutConfig.shortcutKey,
-            this.copySelectionToClipboard,
-            copyShortcutConfig.descriptionKey,
-            copyShortcutConfig.modifiers,
-            false,
-        );
     }
 
     shouldComponentUpdate(nextProps, nextState) {
@@ -275,38 +256,6 @@ class ReportActionsView extends React.Component {
         }
 
         Report.unsubscribeFromReportChannel(this.props.reportID);
-
-        if (this.unsubscribeCopyShortcut) {
-            this.unsubscribeCopyShortcut();
-        }
-    }
-
-    /**
-     * Records the max action on app visibility change event.
-     */
-    onVisibilityChange() {
-        if (!Visibility.isVisible() || this.props.isDrawerOpen) {
-            return;
-        }
-
-        Report.updateLastReadActionID(this.props.reportID);
-    }
-
-    copySelectionToClipboard() {
-        const selectionMarkdown = SelectionScraper.getAsMarkdown();
-
-        Clipboard.setString(selectionMarkdown);
-    }
-
-    /**
-     * Create a unique key for Each Action in the FlatList.
-     * We use a combination of sequenceNumber and clientID in case the clientID are the same - which
-     * shouldn't happen, but might be possible in some rare cases.
-     * @param {Object} item
-     * @return {String}
-     */
-    keyExtractor(item) {
-        return `${item.action.sequenceNumber}${item.action.clientID}`;
     }
 
     fetchData() {
@@ -339,19 +288,6 @@ class ReportActionsView extends React.Component {
     }
 
     /**
-     * Calculates the ideal number of report actions to render in the first render, based on the screen height and on
-     * the height of the smallest report action possible.
-     * @return {Number}
-     */
-    calculateInitialNumToRender() {
-        const minimumReportActionHeight = styles.chatItem.paddingTop + styles.chatItem.paddingBottom
-            + variables.fontSizeNormalHeight;
-        const availableHeight = this.props.windowHeight
-            - (styles.chatFooter.minHeight + variables.contentHeaderHeight);
-        return Math.ceil(availableHeight / minimumReportActionHeight);
-    }
-
-    /**
      * Updates and sorts the report actions by sequence number
      *
      * @param {Array<{sequenceNumber, actionName}>} reportActions
@@ -366,41 +302,6 @@ class ReportActionsView extends React.Component {
             .map((item, index) => ({action: item, index}))
             .value()
             .reverse();
-    }
-
-    /**
-     * Returns true when the report action immediately before the
-     * specified index is a comment made by the same actor who who
-     * is leaving a comment in the action at the specified index.
-     * Also checks to ensure that the comment is not too old to
-     * be considered part of the same comment
-     *
-     * @param {Number} actionIndex - index of the comment item in state to check
-     *
-     * @return {Boolean}
-     */
-    isConsecutiveActionMadeByPreviousActor(actionIndex) {
-        const previousAction = this.sortedReportActions[actionIndex + 1];
-        const currentAction = this.sortedReportActions[actionIndex];
-
-        // It's OK for there to be no previous action, and in that case, false will be returned
-        // so that the comment isn't grouped
-        if (!currentAction || !previousAction) {
-            return false;
-        }
-
-        // Comments are only grouped if they happen within 5 minutes of each other
-        if (currentAction.action.timestamp - previousAction.action.timestamp > 300) {
-            return false;
-        }
-
-        // Do not group if previous or current action was a renamed action
-        if (previousAction.action.actionName === CONST.REPORT.ACTIONS.TYPE.RENAMED
-            || currentAction.action.actionName === CONST.REPORT.ACTIONS.TYPE.RENAMED) {
-            return false;
-        }
-
-        return currentAction.action.actorEmail === previousAction.action.actorEmail;
     }
 
     /**
@@ -526,66 +427,11 @@ class ReportActionsView extends React.Component {
         }
     }
 
-    /**
-     * This function overrides the CellRendererComponent (defaults to a plain View), giving each ReportActionItem a
-     * higher z-index than the one below it. This prevents issues where the ReportActionContextMenu overlapping between
-     * rows is hidden beneath other rows.
-     *
-     * @param {Object} index - The ReportAction item in the FlatList.
-     * @param {Object|Array} style – The default styles of the CellRendererComponent provided by the CellRenderer.
-     * @param {Object} props – All the other Props provided to the CellRendererComponent by default.
-     * @returns {React.Component}
-     */
-    renderCell({item, style, ...props}) {
-        const cellStyle = [
-            style,
-            {zIndex: item.action.sequenceNumber},
-        ];
-        // eslint-disable-next-line react/jsx-props-no-spreading
-        return <View style={cellStyle} {...props} />;
-    }
-
-    /**
-     * Do not move this or make it an anonymous function it is a method
-     * so it will not be recreated each time we render an item
-     *
-     * See: https://reactnative.dev/docs/optimizing-flatlist-configuration#avoid-anonymous-function-on-renderitem
-     *
-     * @param {Object} args
-     * @param {Object} args.item
-     * @param {Number} args.index
-     *
-     * @returns {React.Component}
-     */
-    renderItem({
-        item,
-        index,
-    }) {
-        const shouldDisplayNewIndicator = this.props.report.newMarkerSequenceNumber > 0
-            && item.action.sequenceNumber === this.props.report.newMarkerSequenceNumber;
-        return (
-            <ReportActionItem
-                reportID={this.props.reportID}
-                action={item.action}
-                displayAsGroup={this.isConsecutiveActionMadeByPreviousActor(index)}
-                shouldDisplayNewIndicator={shouldDisplayNewIndicator}
-                isMostRecentIOUReportAction={item.action.sequenceNumber === this.mostRecentIOUReportSequenceNumber}
-                hasOutstandingIOU={this.props.report.hasOutstandingIOU}
-                index={index}
-            />
-        );
-    }
-
     render() {
         // Comments have not loaded at all yet do nothing
         if (!_.size(this.props.reportActions)) {
             return null;
         }
-
-        // Native mobile does not render updates flatlist the changes even though component did update called.
-        // To notify there something changes we can use extraData prop to flatlist
-        const extraData = (!this.props.isDrawerOpen && this.props.isSmallScreenWidth) ? this.props.report.newMarkerSequenceNumber : undefined;
-        const shouldShowReportRecipientLocalTime = ReportUtils.canShowReportRecipientLocalTime(this.props.personalDetails, this.props.report);
 
         return (
             <>
@@ -595,30 +441,16 @@ class ReportActionsView extends React.Component {
                     onClick={this.scrollToBottomAndUpdateLastRead}
                     onClose={this.hideFloatingMessageCounter}
                 />
-                <InvertedFlatList
-                    ref={ReportScrollManager.flatListRef}
-                    data={this.sortedReportActions}
-                    renderItem={this.renderItem}
-                    CellRendererComponent={this.renderCell}
-                    contentContainerStyle={[
-                        styles.chatContentScrollView,
-                        shouldShowReportRecipientLocalTime && styles.pt0,
-                    ]}
-                    keyExtractor={this.keyExtractor}
-                    initialRowHeight={32}
-                    initialNumToRender={this.calculateInitialNumToRender()}
-                    onEndReached={this.loadMoreChats}
-                    onEndReachedThreshold={0.75}
-                    ListFooterComponent={this.props.isLoadingReportActions
-                        ? <ActivityIndicator size="small" color={themeColors.spinner} />
-                        : null}
-                    keyboardShouldPersistTaps="handled"
+                <ReportActionsList
+                    report={this.props.report}
+                    sortedReportActions={this.sortedReportActions}
                     onLayout={this.recordTimeToMeasureItemLayout}
                     onScroll={this.trackScroll}
-                    extraData={extraData}
+                    mostRecentIOUReportSequenceNumber={this.mostRecentIOUReportSequenceNumber}
                 />
                 <PopoverReportActionContextMenu ref={ReportActionContextMenu.contextMenuRef} />
                 <EmojiPicker ref={EmojiPickerAction.emojiPickerRef} />
+                <CopySelectionHelper />
             </>
         );
     }
@@ -632,7 +464,6 @@ export default compose(
     withWindowDimensions,
     withDrawerState,
     withLocalize,
-    withPersonalDetails(),
     withNetwork(),
     withOnyx({
         isLoadingReportData: {
